% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_meta_data.R
\name{add_met_data}
\alias{add_met_data}
\title{Add air quality station meta data to pollutant data}
\usage{
add_met_data(
  df,
  pol,
  date_format,
  date_col,
  id,
  from_here = TRUE,
  csv_path = getwd()
)
}
\arguments{
\item{df}{A data frame with date and pollutant data.}

\item{pol}{Pollutant column name}

\item{date_format}{The format of date data. Default is "\%Y-\%m-\%d \%H:\%M".}

\item{date_col}{Name of column with date data.}

\item{id}{Air quality station id code.}

\item{from_here}{Get local information from your pc.}

\item{csv_path}{Path to save the csv file. Default is working directory.}
}
\value{
csv file
}
\description{
Add air quality station meta data to pollutant data
}
\examples{
\dontrun{
add_met_data(df, "o3", "\%Y-\%m-\%d \%H:\%M", "date", aqs_code)
}
}
