% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_meta_data.R
\name{add_meta_data}
\alias{add_meta_data}
\title{Add air quality station meta data to pollutant data}
\usage{
add_meta_data(
  df,
  pol,
  aqs_meta,
  date_format = "\%Y-\%m-\%d \%H:\%M",
  date_col = "date",
  csv_path = getwd()
)
}
\arguments{
\item{df}{A data frame with date and pollutant data.}

\item{pol}{Pollutant column name}

\item{aqs_meta}{A vector with key value pairs.}

\item{date_format}{The format of date data. Default is "\%Y-\%m-\%d \%H:\%M".}

\item{date_col}{Name of column with date data.}

\item{csv_path}{Path to save the csv file. Default is working directory.}
}
\value{
csv file
}
\description{
Add air quality station meta data to pollutant data
}
\examples{
\dontrun{
library(Dict)
pin_o3 <- readRDS("~/R_tests/pin_meta_test.rds")
pin_meta <- c(
  "Station_id" = "0001",
  "Station_country" = "Brazil",
  "Station_city" = "Sao Paulo",
  "Station_name" = "Pinheiros",
  "Station_timezone" = Sys.timezone(),
  "Timeshift_from_UTC" = as.POSIXlt(Sys.time())$zone
)

add_met_data(pin_o3, "o3", "\%Y-\%m-\%d \%H:\%M", "date", pin_meta,
            csv_path = "~/")
}
}
