% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_meta_template.R
\name{load_meta_template}
\alias{load_meta_template}
\title{Metadata template}
\usage{
load_meta_template()
}
\value{
a character vector with key names
}
\description{
Load the metadata template to fill in the blanks
}
\examples{
\dontrun{
my_meta <- load_meta_template()
}
}
